+++ Shared.hpp
@@ -47,7 +47,7 @@ public:
       uint8_t bitPosition = 0; /**< Bits in c0 (0 to 7), in other words the position of the bit to be predicted (0=MSB) */
       uint32_t c4 = 0; /**< Last 4 whole bytes (buf(4)..buf(1)), packed.  Last byte is bits 0-7. */
       uint32_t c8 = 0; /**< Another 4 bytes (buf(8)..buf(5)) */
-      uint64_t misses{}; //updated by the Predictor, used by SSE stage
+      uint32_t misses{}; //updated by the Predictor, used by SSE stage
 
       BlockType blockType{}; //used by wordModel, recordModel, SSE stage
       uint32_t blockPos{}; //relative position in block, used by many models
